Extract generic assertions from combineReducers
