+++ src/utils/actionTypes.js
@@ -1,0 +1,9 @@
+/**
+ * These are private action types reserved by Redux.
+ * For any unknown actions, you must return the current state.
+ * If the current state is undefined, you must return the initial state.
+ * Do not reference these action types directly in your code.
+ */
+export default {
+  INIT: '@@redux/INIT'
+}
