+++ src/utils/combineReducers.js
@@ -1,19 +1,10 @@
-import { ActionTypes } from '../createStore'
+import ActionTypes from './actionTypes'
+import { assertStateNotUndefined, assertReducerSanity } from './assertions'
 import isPlainObject from './isPlainObject'
 import mapValues from './mapValues'
 import pick from './pick'
 
 /* eslint-disable no-console */
-
-function getUndefinedStateErrorMessage(key, action) {
-  var actionType = action && action.type
-  var actionName = actionType && `"${actionType.toString()}"` || 'an action'
-
-  return (
-    `Reducer "${key}" returned undefined handling ${actionName}. ` +
-    `To ignore an action, you must explicitly return the previous state.`
-  )
-}
 
 function getUnexpectedStateKeyWarningMessage(inputState, outputState, action) {
   var reducerKeys = Object.keys(outputState)
@@ -51,34 +42,6 @@ function getUnexpectedStateKeyWarningMessage(inputState, outputState, action) {
   }
 }
 
-function assertReducerSanity(reducers) {
-  Object.keys(reducers).forEach(key => {
-    var reducer = reducers[key]
-    var initialState = reducer(undefined, { type: ActionTypes.INIT })
-
-    if (typeof initialState === 'undefined') {
-      throw new Error(
-        `Reducer "${key}" returned undefined during initialization. ` +
-        `If the state passed to the reducer is undefined, you must ` +
-        `explicitly return the initial state. The initial state may ` +
-        `not be undefined.`
-      )
-    }
-
-    var type = '@@redux/PROBE_UNKNOWN_ACTION_' + Math.random().toString(36).substring(7).split('').join('.')
-    if (typeof reducer(undefined, { type }) === 'undefined') {
-      throw new Error(
-        `Reducer "${key}" returned undefined when probed with a random type. ` +
-        `Don't try to handle ${ActionTypes.INIT} or other actions in "redux/*" ` +
-        `namespace. They are considered private. Instead, you must return the ` +
-        `current state for any unknown actions, unless it is undefined, ` +
-        `in which case you must return the initial state, regardless of the ` +
-        `action type. The initial state may not be undefined.`
-      )
-    }
-  })
-}
-
 /**
  * Turns an object whose values are different reducer functions, into a single
  * reducer function. It will call every child reducer, and gather their results
@@ -101,7 +64,9 @@ export default function combineReducers(reducers) {
   var sanityError
 
   try {
-    assertReducerSanity(finalReducers)
+    Object.keys(finalReducers).forEach(key => {
+      assertReducerSanity(finalReducers[key], key)
+    })
   } catch (e) {
     sanityError = e
   }
@@ -117,10 +82,7 @@ export default function combineReducers(reducers) {
     var finalState = mapValues(finalReducers, (reducer, key) => {
       var previousStateForKey = state[key]
       var nextStateForKey = reducer(previousStateForKey, action)
-      if (typeof nextStateForKey === 'undefined') {
-        var errorMessage = getUndefinedStateErrorMessage(key, action)
-        throw new Error(errorMessage)
-      }
+      assertStateNotUndefined(nextStateForKey, action, key)
       hasChanged = hasChanged || nextStateForKey !== previousStateForKey
       return nextStateForKey
     })
